+++ sei-ibc-go/modules/apps/transfer/keeper/msg_server_test.go
@@ -2,6 +2,7 @@ package keeper_test
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 
 	"github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
 )
@@ -20,6 +21,17 @@ func (suite *KeeperTestSuite) TestMsgTransfer() {
 			true,
 		},
 		{
+			"bank send enabled for denom",
+			func() {
+				suite.chainA.GetSimApp().BankKeeper.SetParams(suite.chainA.GetContext(),
+					banktypes.Params{
+						SendEnabled: []*banktypes.SendEnabled{{Denom: sdk.DefaultBondDenom, Enabled: true}},
+					},
+				)
+			},
+			true,
+		},
+		{
 			"invalid sender",
 			func() {
 				msg.Sender = "address"
@@ -30,6 +42,17 @@ func (suite *KeeperTestSuite) TestMsgTransfer() {
 			"sender is a blocked address",
 			func() {
 				msg.Sender = suite.chainA.GetSimApp().AccountKeeper.GetModuleAddress(types.ModuleName).String()
+			},
+			false,
+		},
+		{
+			"bank send disabled for denom",
+			func() {
+				suite.chainA.GetSimApp().BankKeeper.SetParams(suite.chainA.GetContext(),
+					banktypes.Params{
+						SendEnabled: []*banktypes.SendEnabled{{Denom: sdk.DefaultBondDenom, Enabled: false}},
+					},
+				)
 			},
 			false,
 		},